+++ lib/view_component/rendering_monkey_patch.rb
@@ -3,7 +3,7 @@
 module ViewComponent
   module RenderingMonkeyPatch # :nodoc:
     def render(options = {}, args = {})
-      if options.respond_to?(:render_in)
+      if options.respond_to?(:render_in) && Rails.version.to_f < 6.1
         self.response_body = options.render_in(self.view_context)
       else
         super