+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # master
 
+* Make `ActionView::Helpers::TagHelper` available in Previews
+
+    def with_html_content
+      render(MyComponent.new) do
+        tag.div do
+          content_tag(:span, "Hello")
+        end
+      end
+    end
+
+    *Sean Doyle*
+
 # v1.14.1
 
 * Fix bug where generator created invalid test code.