+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
     *Joel Hawksley*
 
+* Bypass monkey patch on Rails 6.1 builds.
+
+    *Joel Hawksley*
+    
 * Make `ActionView::Helpers::TagHelper` available in Previews
 
     def with_html_content